add engine metas in /speakers API
