+++ voicevox_engine/metas/MetasStore.py
@@ -1,0 +1,33 @@
+import json
+from pathlib import Path
+
+
+class MetasStore:
+    def __init__(self, engine_speakers_path: Path) -> None:
+        self._engine_speakers_path = engine_speakers_path
+        self._loaded_metas = {
+            folder.name: json.loads((folder / "metas.json").read_text(encoding="utf-8"))
+            for folder in engine_speakers_path.iterdir()
+        }
+
+    def speaker_metas(self, speaker_uuid: str):
+        return self.loaded_metas[speaker_uuid]
+
+    def combine_metas(self, core_metas):
+        """
+        与えられたmetaにエンジンのコア情報を付加して返す
+        core_metas: コアのmetas()が返すJSONのpythonオブジェクト形式
+        """
+
+        return [
+            {**speaker_meta, **self.speaker_metas(speaker_meta["speaker_uuid"])}
+            for speaker_meta in core_metas
+        ]
+
+    @property
+    def engine_speakers_path(self) -> Path:
+        return self._engine_speakers_path
+
+    @property
+    def loaded_metas(self):
+        return self._loaded_metas
