+++ voicevox_engine/metas/__init__.py
@@ -1,0 +1,6 @@
+from . import Metas, MetasStore
+
+__all__ = [
+    "Metas",
+    "MetasStore",
+]
